Refactor and moving stuff around to avoid conflicts
