+++ src/System.CommandLine.CommandHandler/CommandHandlerGenerator.cs
