+++ src/System.CommandLine.Tests/System.CommandLine.Tests.csproj
@@ -16,7 +16,6 @@
   </ItemGroup>
   
   <ItemGroup>
-    <ProjectReference Include="..\..\System.CommandLine.CommandHandler\System.CommandLine.CommandHandler.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false"/>
     <ProjectReference Include="..\System.CommandLine\System.CommandLine.csproj" />
   </ItemGroup>
